requirements.txt
customtkinter
pillow
mysql-connector-python
psutil
paramiko
netmiko
passlib
numpy
regex